Added weight items feature and show item type (kg,pcs), added new method for searchitem in Pos

+++ src/main/java/com/chala/posapp/dto/grn/GrnItemRequest.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 @Data
 public class GrnItemRequest {
     private Long itemId;
-    private Integer qty;
+    private BigDecimal qty;
 
     private BigDecimal costPrice;
     private BigDecimal sellingPrice;
